+++ GR_App.lua
@@ -4,7 +4,6 @@ ns = {}
 
 -- Application Initialization
 -- Uses: AceConsole-3.0, AceEvent-3.0, AceComm-3.0, AceHook-3.0, AceSerializer-3.0
-
 GRADDON = LibStub('AceAddon-3.0'):NewAddon('GuildRecruiter', 'AceConsole-3.0', 'AceEvent-3.0', 'AceHook-3.0', 'AceComm-3.0', 'AceSerializer-3.0')
 local L = LibStub("AceLocale-3.0"):GetLocale('GuildRecruiter')