+++ GR_Settings.lua
@@ -204,7 +204,7 @@ ns.addonSettings = {
                     name = 'Guild Master Settings',
                     type = 'header',
                 },
-                optRememberInvite = {
+                optGMAntiSpamEnable = {
                     order = 2,
                     name = 'Anti guild spam protection.',
                     desc = "Remembers invited players so you don't constantly spam them invites",
@@ -214,7 +214,7 @@ ns.addonSettings = {
                     set = function(_, val) ns.dbGlobal.guildInfo.antiSpam = val end,
                     get = function() return ns.dbGlobal.guildInfo.antiSpam end,
                 },
-                optReInvite = {
+                optGMAntiSpamInterval = {
                     order = 3,
                     name = 'Reinvite players after:',
                     desc = 'Number of days before resetting invite status.',
@@ -224,14 +224,19 @@ ns.addonSettings = {
                     disabled = function() return not ns.isGuildLeader end,
                     values = function()
                         return {
-                            [1] = '1 day',
-                            [3] = '3 days',
-                            [5] = '5 days',
                             [7] = '7 days',
+                            [14] = '14 days',
+                            [30] = '30 days (1 month)',
+                            [190] = '190 days (3 months)',
+                            [380] = '380 days (6 months)',
                         }
                     end,
                     set = function(_, val) ns.dbGlobal.guildInfo.reinviteAfter = tonumber(val) end,
-                    get = function() return ns.dbGlobal.guildInfo.reinviteAfter end,
+                    get = function()
+                        if ns.dbGlobal.guildInfo.reinviteAfter and ns.dbGlobal.guildInfo.reinviteAfter < 7 then
+                            ns.dbGlobal.guildInfo.reinviteAfter = 7
+                        end
+                        return ns.dbGlobal.guildInfo.reinviteAfter end,
                 },
                 optOnlyGMGreeting = {
                     order = 4,
@@ -505,7 +510,7 @@ ns.addonSettings = {
                     get = function() return ns.settings.showContext end,
                 },
                 optShowInvite = {
-                    order = 2,
+                    order = 3,
                     name = 'Show your whisper when sending invite messages.',
                     desc = 'This will show or hide whisper messages to recruits, suggest going to social and turn on in-line whispers for best results.',
                     type = 'toggle',
@@ -514,13 +519,50 @@ ns.addonSettings = {
                     get = function() return ns.settings.showWhispers end,
                 },
                 optShowInviteMsg = {
-                    order = 3,
+                    order = 4,
                     name = ns.code:cText('FF00FF00', 'NOTE: You must reload your UI to take effect (/rl).'),
                     type = 'description',
                     fontSize = 'medium',
                 },
+                optGMAntiSpamEnable = {
+                    order = 5,
+                    name = 'Anti guild spam protection.',
+                    desc = "Remembers invited players so you don't constantly spam them invites",
+                    type = 'toggle',
+                    disabled = function() return ns.dbGlobal.guildInfo.antiSpam end,
+                    width = 1.5,
+                    set = function(_, val) ns.settings.antiSpam = val end,
+                    get = function()
+                        if ns.dbGlobal.guildInfo.antiSpam then
+                            return ns.dbGlobal.guildInfo.antiSpam
+                        else return ns.settings.antiSpam end
+                    end,
+                },
+                optGMAntiSpamInterval = {
+                    order = 6,
+                    name = 'Reinvite players after:',
+                    desc = 'Number of days before resetting invite status.',
+                    type = 'select',
+                    style = 'dropdown',
+                    width = 1,
+                    disabled = function() return ns.dbGlobal.guildInfo.antiSpam end,
+                    values = function()
+                        return {
+                            [7] = '7 days',
+                            [14] = '14 days',
+                            [30] = '30 days (1 month)',
+                            [190] = '190 days (3 months)',
+                            [380] = '380 days (6 months)',
+                        }
+                    end,
+                    set = function(_, val) ns.settings.reinviteAfter = tonumber(val) end,
+                    get = function()
+                        if ns.dbGlobal.guildInfo.antiSpam then return ns.dbGlobal.guildInfo.reinviteAfter
+                        else return ns.settings.reinviteAfter end
+                    end,
+                },
                 optShowPersonalGreeting = {
-                    order = 4,
+                    order = 7,
                     name = 'Enable/Disable whispered greeting message to new guild member.',
                     desc = 'This message will be in a '..ns.code:cText('FFFFFF00', 'whisper')..' form to the player upon joining.',
                     type = 'toggle',
@@ -530,14 +572,14 @@ ns.addonSettings = {
                     get = function() return ns.dbGlobal.guildInfo.greeting or ns.settings.sendGreeting end,
                 },
                 optShowPersonalGM = {
-                    order = 5,
+                    order = 8,
                     name = ns.code:cText('FF00FF00', 'NOTE: The GM has forced this option on.'),
                     hidden = function() return not ns.dbGlobal.guildInfo.greeting end,
                     type = 'description',
                     fontSize = 'medium',
                 },
                 optPersonalGreetingMsg = {
-                    order = 6,
+                    order = 9,
                     name = 'Actual greeting message',
                     desc = 'This is the message that will be '..ns.code:cText('FFFFFF00', 'whispered')..' to the player after joining.',
                     type = 'input',
@@ -555,10 +597,10 @@ ns.addonSettings = {
                 Header2 = {
                     name = '',
                     type = 'header',
-                    order = 7
+                    order = 10
                 },
                 PreviewCount = {
-                    order = 8,
+                    order = 11,
                     name = function()
                         local gi = ns.dbGlobal.guildInfo
                         local playerNameFound, count = MessageLength((ns.settings.greetingMsg or gi.greetingMsg or ''))
@@ -571,12 +613,12 @@ ns.addonSettings = {
                     fontSize = 'medium'
                 },
                 Header3 = {
-                    order = 9,
+                    order = 12,
                     name = '',
                     type = 'header',
                 },
                 optShowPersonalWelcome = {
-                    order = 10,
+                    order = 13,
                     name = 'Enable/Disable personal welcome message.',
                     desc = 'Enable/Disable sending of a personalized welcome message to '..ns.code:cText('FFFFFF00', 'guild chat')..' after a player joins.',
                     type = 'toggle',
@@ -586,7 +628,7 @@ ns.addonSettings = {
                     get = function() return ns.settings.sendWelcome end,
                 },
                 optPersonalWelcomeMsg = {
-                    order = 11,
+                    order = 14,
                     name = 'Actual message sent to player after joining.',
                     desc = 'This message will be sent to '..ns.code:cText('FFFFFF00', 'guild chat')..' to the player after joining.',
                     type = 'input',
@@ -603,14 +645,14 @@ ns.addonSettings = {
                     end,
                 },
                 optMsgNote = {
-                    order = 12,
+                    order = 15,
                     name = ns.code:cText('FFFFFF00', 'NOTE: ')..'Disabled options are controlled by the Guild Master.',
                     type = 'description',
                     width = 'full',
                     fontSize = 'medium'
                 },
                 optScanInterval = {
-                    order = 13,
+                    order = 16,
                     name = 'Time to wait between scans (default recommended).',
                     desc = 'WoW requires a cooldown period between /who scans, this is the time that the system will wait between scans.',
                     type = 'input',
@@ -622,7 +664,7 @@ ns.addonSettings = {
                     get = function() return tostring(ns.settings.scanWaitTime) end,
                 },
                 optWhoNote = {
-                    order = 14,
+                    order = 17,
                     name = ns.code:cText('FFFFFF00', 'NOTE: ')..'6 seconds seems to give best results, shorter time yields less results.',
                     type = 'description',
                     width = 'full',
@@ -1132,12 +1174,12 @@ ns.addonSettings = {
                 },
                 blRemoveButton = {
                     name = 'Toggle Selected Black List Entries',
-                    desc = 'Black List entries marked for deletion will be perinately removed 30 days after marked.  During this time, the addon will ignore the selected Black List entries.',
+                    desc = 'Black List entries marked for deletion will be permanently removed 30 days after marked.  During this time, the addon will ignore the selected Black List entries.',
                     type = 'execute',
                     width = 'full',
                     order = 2,
                     func = function()
-                        for _,r in pairs(ns.dbBL) do
+                        for _,r in pairs(ns.tblBlackList) do
                             if r.selected and not r.markedForDelete then
                                 r.markedForDelete = true
                                 r.expirationDate = C_DateAndTime.GetServerTimeLocal() + (30 * SECONDS_IN_A_DAY)
@@ -1156,7 +1198,7 @@ ns.addonSettings = {
                     width = 'full',
                     values = function()
                         local tbl = {}
-                        for k, r in pairs(ns.dbBL or {}) do
+                        for k, r in pairs(ns.tblBlackList or {}) do
                             local name = k
                             name = r.markedForDelete and ns.code:cText('FFFF0000', name) or ns.code:cText('FF00FF00', name)
                             tbl[k] = (name..': '..(r.reason or 'Unknown'))
@@ -1164,8 +1206,8 @@ ns.addonSettings = {
 
                         return tbl
                     end,
-                    set = function(_, key, val) ns.dbBL[key].selected = val end,
-                    get = function(_, key) return ns.dbBL[key].selected or false end,
+                    set = function(_, key, val) ns.tblBlackList[key].selected = val end,
+                    get = function(_, key) return ns.tblBlackList[key].selected or false end,
                 }
             }
         },